First version of refactor project, remove global React from every component, edit some context structure, add husky to project
