+++ index.html
@@ -23,5 +23,5 @@
 <div id='cursor'></div>
 <div id='modals'></div>
 </body>
-<script type='module' src='/src/main.tsx'></script>
+<script defer type='module' src='/src/main.tsx'></script>
 </html>
