+++ src/App.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import { useEffect } from 'react';
 import Routing from '@/components/Routing';
 import Cursor from '@/components/Cursor';
 import DevToolsModal from '@/components/DevToolsModal';
