+++ src/components/BasicPage/BasicPage.tsx
@@ -1,7 +1,7 @@
-import React, { PropsWithChildren } from 'react';
+import { FC, PropsWithChildren } from 'react';
 import cl from './BasicPage.module.scss';
 
-const BasicPage: React.FC<PropsWithChildren> = ({ children }) => {
+const BasicPage: FC<PropsWithChildren> = ({ children }) => {
   return <main className={cl.basicPage}>{children}</main>;
 };
