+++ src/hooks/useToggle.ts
@@ -1,8 +1,9 @@
 import { useCallback, useState } from 'react';
 
-export default (
+export default function useToggle(
   initialState: boolean,
-): [typeof initialState, (newState?: typeof initialState) => void] => {
+): [typeof initialState, (newState?: typeof initialState) => void];
+export default function useToggle(initialState) {
   const [state, setState] = useState<typeof initialState>(initialState);
 
   const changeState = useCallback(
