+++ src/index.scss
@@ -100,19 +100,5 @@ input {
 }
 
 @mixin underlineText {
-  position: relative;
-
-  &:after {
-    --deco-height: 0.3125em;
-    content: '';
-    position: absolute;
-    left: 0;
-    right: 0;
-    bottom: calc(var(--deco-height) * -0.625);
-    height: var(--deco-height);
-    background-size: auto 110%;
-    background-repeat: repeat-x;
-    background-position: 0;
-    background-image: url('!/underline.svg');
-  }
+  border-bottom: 2px solid #00A080;
 }
