+++ src/pages/AboutPage/AboutPage.module.scss
@@ -44,6 +44,7 @@
       }
 
       span {
+        @include underlineText;
         @include circlesText;
         font-family: var(--font-secondary);
         text-shadow: 0 0 20px var(--text-green);
