+++ src/pages/NotFoundPage/NotFoundPage.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import PageContainer from '@/components/PageContainer';
 
 const NotFoundPage = () => {
