+++ src/types/DispatchAction.ts
@@ -1,0 +1,4 @@
+export interface DispatchAction<T=any> {
+  type: string;
+  payload: T;
+}
