+++ src/utils/TypeOf.ts
@@ -1,0 +1,7 @@
+export function TypeOf(value) {
+  return Object.prototype.toString
+    .call(value)
+    .split(' ')[1]
+    .slice(0, -1)
+    .toLowerCase();
+}
