+++ src/utils/createContextDefaultState.ts
@@ -1,0 +1,8 @@
+import { ContextState } from '@/types/ContextState';
+
+export const createContextDefaultState = <T = any>(state: T): ContextState<T> => ({
+  state,
+  dispatch: function(action) {
+    return this.state;
+  },
+});
