+++ tsconfig.json
@@ -22,6 +22,7 @@
     "types": [
       "vite/client"
     ],
+    "noImplicitAny": false,
     "baseUrl": ".",
     "paths": {
       "@/*": [
